More changes for more items support

+++ Comet-Server/src/main/java/com/cometproject/server/game/items/rares/LimitedEditionItem.java
@@ -3,17 +3,17 @@ package com.cometproject.server.game.items.rares;
 public class LimitedEditionItem {
     public static final LimitedEditionItem NONE = new LimitedEditionItem(0, 0, 0);
 
-    private int itemId;
+    private long itemId;
     private int limitedRare;
     private int limitedRareTotal;
 
-    public LimitedEditionItem(int itemId, int limitedRare, int limitedRareTotal) {
+    public LimitedEditionItem(long itemId, int limitedRare, int limitedRareTotal) {
         this.itemId = itemId;
         this.limitedRare = limitedRare;
         this.limitedRareTotal = limitedRareTotal;
     }
 
-    public int getItemId() {
+    public long getItemId() {
         return itemId;
     }
