+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/GenericFloorItem.java
@@ -8,7 +8,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public class GenericFloorItem extends RoomItemFloor {
-    public GenericFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public GenericFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
