+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/GenericWallItem.java
@@ -7,7 +7,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public final class GenericWallItem extends RoomItemWall {
-    public GenericWallItem(int id, int itemId, Room room, int owner, String position, String data) {
+    public GenericWallItem(long id, int itemId, Room room, int owner, String position, String data) {
         super(id, itemId, room, owner, position, data);
     }
