+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/AdjustableHeightSeatFloorItem.java
@@ -5,7 +5,7 @@ import org.apache.commons.lang.StringUtils;
 
 
 public class AdjustableHeightSeatFloorItem extends SeatFloorItem {
-    public AdjustableHeightSeatFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public AdjustableHeightSeatFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         if (this.getExtraData().isEmpty()) {
