+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/AlertFloorItem.java
@@ -7,7 +7,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public class AlertFloorItem extends RoomItemFloor {
-    public AlertFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public AlertFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
