+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/DiceFloorItem.java
@@ -12,7 +12,7 @@ public class DiceFloorItem extends RoomItemFloor {
     private boolean isInUse = false;
     private int rigNumber = -1;
 
-    public DiceFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public DiceFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
