+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/GateFloorItem.java
@@ -9,7 +9,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public class GateFloorItem extends RoomItemFloor {
-    public GateFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public GateFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
