+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/LoveLockFloorItem.java
@@ -13,7 +13,7 @@ public class LoveLockFloorItem extends RoomItemFloor {
     private int leftEntity = 0;
     private int rightEntity = 0;
 
-    public LoveLockFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public LoveLockFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
