+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/MagicStackFloorItem.java
@@ -9,7 +9,7 @@ import java.text.DecimalFormat;
 public class MagicStackFloorItem extends RoomItemFloor {
     private double magicHeight = 0d;
 
-    public MagicStackFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public MagicStackFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
