+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/OneWayGateFloorItem.java
@@ -11,7 +11,7 @@ public class OneWayGateFloorItem extends RoomItemFloor {
     private boolean isInUse = false;
     private GenericEntity interactingEntity;
 
-    public OneWayGateFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public OneWayGateFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
