+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/PressurePlateSeatFloorItem.java
@@ -4,7 +4,7 @@ import com.cometproject.server.game.rooms.objects.entities.GenericEntity;
 import com.cometproject.server.game.rooms.types.Room;
 
 public class PressurePlateSeatFloorItem extends SeatFloorItem {
-    public PressurePlateSeatFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public PressurePlateSeatFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
