+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/PrivateChatFloorItem.java
@@ -13,7 +13,7 @@ public class PrivateChatFloorItem extends GenericFloorItem {
 
     private List<PlayerEntity> entities = Lists.newArrayList();
 
-    public PrivateChatFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public PrivateChatFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
