+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/SoundMachineFloorItem.java
@@ -29,7 +29,7 @@ public class SoundMachineFloorItem extends RoomItemFloor implements Stateable {
 
     private boolean pendingPlay = false;
 
-    public SoundMachineFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public SoundMachineFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         if (data.startsWith("##jukeboxOn[")) {
