+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/banzai/BanzaiTeleporterFloorItem.java
@@ -20,7 +20,7 @@ public class BanzaiTeleporterFloorItem extends RoomItemFloor {
     private GenericEntity entity;
     private RoomItemFloor floorItem;
 
-    public BanzaiTeleporterFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public BanzaiTeleporterFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
