+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/banzai/BanzaiTileFloorItem.java
@@ -17,7 +17,7 @@ public class BanzaiTileFloorItem extends RoomItemFloor {
     private GameTeam gameTeam = GameTeam.NONE;
     private int points = 0;
 
-    public BanzaiTileFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public BanzaiTileFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         this.setExtraData("0");
