+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/banzai/BanzaiTimerFloorItem.java
@@ -11,7 +11,7 @@ import org.apache.commons.lang.StringUtils;
 public class BanzaiTimerFloorItem extends RoomItemFloor {
     private String lastTime;
 
-    public BanzaiTimerFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public BanzaiTimerFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
