+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/football/FootballScoreFloorItem.java
@@ -8,7 +8,7 @@ import com.cometproject.server.game.rooms.types.components.games.GameTeam;
 public class FootballScoreFloorItem extends RoomItemFloor {
     private GameTeam gameTeam;
 
-    public FootballScoreFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public FootballScoreFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         this.setExtraData("0");
