+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/groups/GroupFloorItem.java
@@ -12,7 +12,7 @@ import org.apache.commons.lang.StringUtils;
 public class GroupFloorItem extends RoomItemFloor {
     private int groupId;
 
-    public GroupFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public GroupFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         if (!StringUtils.isNumeric(data))
