+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/groups/GroupGateFloorItem.java
@@ -7,7 +7,7 @@ import com.cometproject.server.network.messages.outgoing.room.items.UpdateFloorI
 
 
 public class GroupGateFloorItem extends GroupFloorItem {
-    public GroupGateFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public GroupGateFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
