+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/snowboarding/SnowboardSlopeFloorItem.java
@@ -5,7 +5,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public class SnowboardSlopeFloorItem extends AdjustableHeightFloorItem {
-    public SnowboardSlopeFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public SnowboardSlopeFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
 }
