+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/summer/SummerShowerFloorItem.java
@@ -6,7 +6,7 @@ import com.cometproject.server.game.rooms.types.Room;
 
 
 public class SummerShowerFloorItem extends RoomItemFloor {
-    public SummerShowerFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public SummerShowerFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
