+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/totem/TotemHeadFloorItem.java
@@ -3,7 +3,7 @@ package com.cometproject.server.game.rooms.objects.items.types.floor.totem;
 import com.cometproject.server.game.rooms.types.Room;
 
 public class TotemHeadFloorItem extends TotemPartFloorItem {
-    public TotemHeadFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public TotemHeadFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
 }
