+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/wired/WiredFloorItem.java
@@ -39,7 +39,7 @@ public abstract class WiredFloorItem extends RoomItemFloor implements WiredItemS
      * @param rotation The orientation of the item
      * @param data     The JSON object associated with this item
      */
-    public WiredFloorItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public WiredFloorItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
 
         if (!this.getExtraData().startsWith("{")) {
