+++ Comet-Server/src/main/java/com/cometproject/server/game/rooms/objects/items/types/floor/wired/actions/WiredActionBotGiveHandItem.java
@@ -25,7 +25,7 @@ public class WiredActionBotGiveHandItem extends WiredActionItem {
      * @param rotation The orientation of the item
      * @param data     The JSON object associated with this item
      */
-    public WiredActionBotGiveHandItem(int id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
+    public WiredActionBotGiveHandItem(long id, int itemId, Room room, int owner, int x, int y, double z, int rotation, String data) {
         super(id, itemId, room, owner, x, y, z, rotation, data);
     }
